# api-gateway-project/api/docs/swagger.yaml
definitions:
...
  models.CreatePost:
    properties:
      category:
        type: string
      content:
        type: string
      dislikes:
        type: integer
      id:
        type: string
      likes:
        type: integer
      title:
        type: string
      user_id:
        type: string
      views:
        type: integer
    type: object
...
  models.Post:
    properties:
      category:
        type: string
      content:
        type: string
      dislikes:
        type: integer
      id:
        type: string
      likes:
        type: integer
      title:
        type: string
      user_id:
        type: string
      views:
        type: integer
    type: object
...
  models.StandardErrorModel:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
host: localhost:7007
info:
  contact: {}
  description: api-gateway
  title: EXAM_TASK_4_API
  version: "1.0"
paths:
...
  /v1/post/{id}:
    get:
      consumes:
      - application/json
      description: Api for GetPost
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreatePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: GetPost
      tags:
      - post
  /v1/post/create:
    post:
      consumes:
      - application/json
      description: Api for CreatePost
      parameters:
      - description: createUserModel
        in: body
        name: Post
        required: true
        schema:
          $ref: '#/definitions/models.CreatePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreatePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: CreatePost
      tags:
      - post
  /v1/post/delete:
    delete:
      consumes:
      - application/json
      description: Api for DeletePost
      parameters:
      - description: ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Post'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: DeletePost
      tags:
      - post
  /v1/post/update:
    put:
      consumes:
      - application/json
      description: Api for UpdatePost
      parameters:
      - description: createUserModel
        in: body
        name: Post
        required: true
        schema:
          $ref: '#/definitions/models.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreatePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: UpdatePost
      tags:
      - post
  /v1/posts:
    get:
      consumes:
      - application/json
      description: Api for GetAllPosts
      parameters:
      - description: User PAGES
        in: query
        name: page
        required: true
        type: string
      - description: User LIMIT
        in: query
        name: limit
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreatePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: GetAllPosts
      tags:
      - post
  /v1/postsbyuserid/{id}:
    get:
      consumes:
      - application/json
      description: Api for GetPostsByUserId
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CreatePost'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.StandardErrorModel'
      security:
      - ApiKeyAuth: []
      summary: GetPostsByUserId
      tags:
      - post
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"